+++ src/App.scss
@@ -2,19 +2,4 @@ iframe {
   display: none;
 }
 
-.Person {
-  width: fit-content;
-  margin-bottom: 16px;
-  padding: 8px;
-  border: 1px solid #000;
-  border-radius: 8px;
 
-  &__name {
-    margin-top: 8px;
-  }
-
-  &__age {
-    color: #00f;
-    font-weight: bold;
-  }
-}
