+++ src/components/Person/Person.scss
@@ -1,0 +1,16 @@
+.Person {
+  width: fit-content;
+  margin-bottom: 16px;
+  padding: 8px;
+  border: 1px solid #000;
+  border-radius: 8px;
+
+  &__name {
+    margin-top: 8px;
+  }
+
+  &__age {
+    color: #00f;
+    font-weight: bold;
+  }
+}
